Finalizando

+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/PlanSalud.java
@@ -6,26 +6,25 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-
 @Entity
 public class PlanSalud {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer idPlanSalud;
-	
+
 	@Column
 	private String nombre;
-	
+
 	@Column
 	private Double descuento;
-	
+
 	@Column
-	private Double precioDelPlan;
-	
+	private Float precioDelPlan;
+
 	public PlanSalud() {
-		
-	}
+
+	} 
 
 	public Integer getIdPlanSalud() {
 		return idPlanSalud;
@@ -51,15 +50,12 @@ public class PlanSalud {
 		this.descuento = descuento;
 	}
 
-	public Double getPrecioDelPlan() {
+	public Float getPrecioDelPlan() {
 		return precioDelPlan;
 	}
 
-	public void setPrecioDelPlan(Double precioDelPlan) {
+	public void setPrecioDelPlan(Float precioDelPlan) {
 		this.precioDelPlan = precioDelPlan;
 	}
-	
-	
-
 
 }
